Fix FILES variable
The list of files and directories that are placed in a package. The PACKAGES variable lists the packages generated by a recipe.

This commit includes the start-stop script in the resulting package.

Prior to this commit, the script was installed and executed upon boot just fine. This commit is just to avoid the following QA error

```
Files/directories were installed but not shipped in any package [installed-vs-shipped]
```